+++ Nos/Service/CurrentUser.swift
@@ -177,7 +177,7 @@ import Dependencies
     }
 
     /// Subscribes to relays for important events concerning the current user like their latest contact list, 
-    /// notifications, reports, mutes, zaps, etc.
+    /// notifications, reports, mutes, zaps, lists etc.
     @MainActor func subscribe() async {
         guard let key = publicKeyHex, let author else {
             return
@@ -191,11 +191,16 @@ import Dependencies
             await relayService.requestContactList(for: key, since: author.lastUpdatedContactList)
         )
         
+        // Always request the user's lists
+        subscriptions.append(
+            await relayService.requestAuthorLists(for: key, since: nil)
+        )
+        
         // Subscribe to important events we may not get incidentally while browsing the feed
         let latestReceivedEvent = try? viewContext.fetch(Event.lastReceived(for: author)).first
         let importantEventsFilter = Filter(
             authorKeys: [key],
-            kinds: [.mute, .delete, .report, .contactList, .zapRequest],
+            kinds: [.mute, .delete, .report, .contactList, .zapRequest, .followSet],
             limit: 100,
             since: latestReceivedEvent?.receivedAt,
             keepSubscriptionOpen: true